Moved service initialization to lib.rs
